Raminha

+++ Backend1/Bancosol/src/main/java/uma/grupo13/bancosol/dao/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import uma.grupo13.bancosol.entity.UsuarioEntity;
 
-public interface UserRepository extends JpaRepository<UsuarioEntity, Long> {
+public interface UserRepository extends JpaRepository<UsuarioEntity, Integer> {
 
     @Query("select u from UsuarioEntity u where u.usuario = :user and u.contrasena = :pwd")
     public UsuarioEntity autheticate (@Param("user")String username, @Param("pwd") String password);
